Replace United RPMS by RPM Fusion
I thought that United RPMS was better (when installing the alpha/beta
version of Fedora, the United RPMS repo is already ready while RPM
Fusion takes more time to be available for a new Fedora version,
sometimes it is not even ready when the new Fedora stable version is
released). But recently unstable versions of GStreamer (latest versions
from git) have been pushed to United RPMS, while some other GStreamer
packages were still at the previous stable version, which resulted in
bugs in the program that I was developing (cosy-camera-recorder). In
short, the United RPMS packagers are not good packagers, it's stupid to
take the latest versions from Git, for a stable distro branch it's much
better to take stable versions, and for a group of related modules, take
the same versions!

So, back to RPM Fusion.